backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.0

# Authentication
python-jose[cryptography]==3.3.0
python-keycloak==3.9.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# S3 Storage
boto3==1.34.34

# Image Processing & Metadata
Pillow==12.1.1
pillow-heif==1.2.1
exifread==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
